fix: satisfy pkg-file-count limits for durable session contract
Move durable session apisurface mappers into pkg/apisurface/factorysession
and consolidate factorysessionexecution source files to stay within the
15-file package cap. Updates fixture test paths and api-relevant-files docs.
