+++ pkg/apisurface/factorysession/factory_session_fake_consumer_test.go
@@ -1,4 +1,4 @@
-package apisurface_test
+package factorysession_test
 
 import (
 	"context"
@@ -6,13 +6,13 @@ import (
 	"testing"
 
 	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
-	"github.com/portpowered/infinite-you/pkg/apisurface"
+	"github.com/portpowered/infinite-you/pkg/apisurface/factorysession"
 	"github.com/portpowered/infinite-you/pkg/factorysessionexecution"
 )
 
 // pkgmaintcheck:ignore-cyclomatic-complexity this consumer test keeps fake-service projection assertions together across apisurface mappers.
 func TestFakeServiceConsumer_ProjectsFixtureThroughApisurfaceMappers(t *testing.T) {
-	fixturesPath := filepath.Join("..", "api", "testdata", "durable-session-contract-fixtures.json")
+	fixturesPath := filepath.Join("..", "..", "api", "testdata", "durable-session-contract-fixtures.json")
 	service, err := factorysessionexecution.NewFakeServiceFromContractFixtures(fixturesPath)
 	if err != nil {
 		t.Fatalf("NewFakeServiceFromContractFixtures: %v", err)
@@ -23,7 +23,7 @@ func TestFakeServiceConsumer_ProjectsFixtureThroughApisurfaceMappers(t *testing.
 	if !ok {
 		t.Fatal("missing executionRequest fixture")
 	}
-	request, err := apisurface.StartRequestFromAPI(decodeExecutionRequest(t, executionRequest))
+	request, err := factorysession.StartRequestFromAPI(decodeExecutionRequest(t, executionRequest))
 	if err != nil {
 		t.Fatalf("StartRequestFromAPI: %v", err)
 	}
@@ -32,7 +32,7 @@ func TestFakeServiceConsumer_ProjectsFixtureThroughApisurfaceMappers(t *testing.
 	if err != nil {
 		t.Fatalf("StartAsync: %v", err)
 	}
-	mappedStart := apisurface.AsyncStartResponseToAPI(started)
+	mappedStart := factorysession.AsyncStartResponseToAPI(started)
 	if mappedStart.SessionId != "dur-sess-petri-success-001" {
 		t.Fatalf("sessionId = %q", mappedStart.SessionId)
 	}
@@ -44,7 +44,7 @@ func TestFakeServiceConsumer_ProjectsFixtureThroughApisurfaceMappers(t *testing.
 	if err != nil {
 		t.Fatalf("GetSession: %v", err)
 	}
-	mappedRead := apisurface.SessionReadResponseToAPI(read)
+	mappedRead := factorysession.SessionReadResponseToAPI(read)
 	if mappedRead.Status != factoryapi.FactorySessionDurableLifecycleStatusSucceeded {
 		t.Fatalf("read status = %q", mappedRead.Status)
 	}
@@ -53,7 +53,7 @@ func TestFakeServiceConsumer_ProjectsFixtureThroughApisurfaceMappers(t *testing.
 	if err != nil {
 		t.Fatalf("ListDispatches: %v", err)
 	}
-	mappedDispatches := apisurface.ListDispatchesResponseToAPI(dispatches)
+	mappedDispatches := factorysession.ListDispatchesResponseToAPI(dispatches)
 	if len(mappedDispatches.Dispatches) != 1 || mappedDispatches.Dispatches[0].Id != "disp-petri-success-001" {
 		t.Fatalf("dispatches = %#v", mappedDispatches.Dispatches)
 	}
@@ -62,7 +62,7 @@ func TestFakeServiceConsumer_ProjectsFixtureThroughApisurfaceMappers(t *testing.
 	if err != nil {
 		t.Fatalf("GetResult: %v", err)
 	}
-	mappedResult := apisurface.ResultResponseToAPI(result)
+	mappedResult := factorysession.ResultResponseToAPI(result)
 	if mappedResult.ResultStatus != factoryapi.FactorySessionResultStatusFinal {
 		t.Fatalf("resultStatus = %q, want FINAL", mappedResult.ResultStatus)
 	}
@@ -74,7 +74,7 @@ func TestFakeServiceConsumer_ProjectsFixtureThroughApisurfaceMappers(t *testing.
 	if err != nil {
 		t.Fatalf("ReadEvents: %v", err)
 	}
-	stream := apisurface.FactoryEventStreamFromReadResult(events)
+	stream := factorysession.FactoryEventStreamFromReadResult(events)
 	if len(stream.History) == 0 {
 		t.Fatal("terminal events missing")
 	}
