+++ pkg/apisurface/factorysession/factory_session_listing.go
@@ -1,4 +1,4 @@
-package apisurface
+package factorysession
 
 import (
 	"strings"
