+++ pkg/apisurface/factorysession/factory_session_projection.go
@@ -1,4 +1,4 @@
-package apisurface
+package factorysession
 
 import (
 	"encoding/json"
