+++ pkg/apisurface/factorysession/helpers.go
@@ -1,0 +1,12 @@
+package factorysession
+
+func derefString(value *string) string {
+	if value == nil {
+		return ""
+	}
+	return *value
+}
+
+func deref(value *string) string {
+	return derefString(value)
+}
