+++ pkg/factorysessionexecution/service.go
@@ -1,6 +1,9 @@
 package factorysessionexecution
 
-import "context"
+import (
+	"context"
+	"fmt"
+)
 
 // Service is the shared durable factory-session execution contract consumed by
 // API, CLI, MCP, and UI transports. Live-session open and invocation remain on
@@ -23,4 +26,33 @@ type Service interface {
 	GetArtifact(ctx context.Context, sessionID, artifactID string) (ArtifactDetail, error)
 	ReadEvents(ctx context.Context, sessionID string, req EventReconnectRequest) (EventReadResult, error)
 	ListSessions(ctx context.Context, req ListSessionsRequest) (ListSessionsResult, error)
+}
+
+// InspectionLinksForSession builds API-relative inspection links for one durable session.
+func InspectionLinksForSession(sessionID string, includeEvents bool) InspectionLinks {
+	base := fmt.Sprintf("/factory-sessions/%s", sessionID)
+	links := InspectionLinks{
+		Session:    base,
+		Status:     base,
+		Results:    base + "/results",
+		Dispatches: base + "/dispatches",
+		Artifacts:  base + "/artifacts",
+	}
+	if includeEvents {
+		links.Events = base + "/events"
+	}
+	return links
+}
+
+// LifecycleControlLinksForSession builds post-control inspection links for one durable session.
+func LifecycleControlLinksForSession(sessionID string, includeEvents bool) LifecycleControlLinks {
+	inspection := InspectionLinksForSession(sessionID, includeEvents)
+	return LifecycleControlLinks{
+		Session:    inspection.Session,
+		Status:     inspection.Status,
+		Results:    inspection.Results,
+		Dispatches: inspection.Dispatches,
+		Artifacts:  inspection.Artifacts,
+		Events:     inspection.Events,
+	}
 }
